requirements.txt
flask
ldap3
pyodbc
flask-talisman
uuid
cryptography
python-dotenv
requests
pytz